s2mpg1x: i2c: Make i2c remove callbacks return void
The upstream commit ed5c2f5fd10d ("i2c: Make remove callback return
void") changes the return value for the i2c_device remove callback to be
void. So we need to update all of these drivers accordingly. The reason
was because the return value wasn't really being used which is
consistent with this driver as well.

Change-Id: I919eb8fa497edad35ad6c27d0c8dbb8c261f6cc6
Signed-off-by: Aarqw12 <lcockx@protonmail.com>